requirements.txt
pyautogen>=0.2.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0